remove one subtraction and fix bug in 32 bit cbrt

+++ C/oroots.h
@@ -52,7 +52,7 @@ inline unsigned int leading_zeros_ull(const unsigned long long *val) {
 	return 0;
 }
 
-inline unsigned int leading_zeros_ui(const unsigned long long *val) {
+inline unsigned int leading_zeros_ui(const unsigned int *val) {
 	unsigned int length = sizeof(*val) << 3;
 	for (unsigned int i = length; i > 0; --i) {
 		if (*val & (1 << (i - 1))) {
